requirements.txt
python-dotenv
watchdog
redis
rq
sqlalchemy
mysql-connector-python
apscheduler